server/requirements.txt
flask
mysql-connector-python
pandas
opencv-python
boto3